+++ src/index.ts
@@ -1,1 +1,2 @@
 export { CustomQueryBuilder, CustomQueryBuilderError } from './CustomQueryBuilder';
+export type { Relation } from './CustomQueryBuilder';